requirements.txt
flask
flask-bootstrap
flask-sqlalchemy
flask-wtf
wtforms
flask-bcrypt
flask-login
PyJWT
flask-mail